create dummy User and remove origin Post #26
