--- src/Components/Post/testPost/index.scss
@@ -1,82 +1,0 @@
-.post {
-  margin: 1%;
-  padding: 1%;
-  display: flex;
-  flex-grow: 3;
-  max-width: 350px;
-  height: 550px;
-  flex-direction: column;
-  align-items: center;
-  font-family: 'Montserrat', sans-serif;
-  position: relative;
-
-  &:hover {
-    background-color: #f2f2f2;
-  }
-
-  &__hover {
-    z-index: 1;
-    &__icon {
-      position: absolute;
-      .bi {
-        cursor: pointer;
-        width: 50px;
-        height: auto;
-        z-index: 2;
-        border: 1px solid #d6c4c3;
-        border-radius: 50%;
-        padding: 10px;
-        margin-right: 15px;
-        color: #d2b2d9;
-        background-color: white;
-        &.bi-brush {
-          color: #f59b47;
-        }
-      }
-    }
-  }
-
-  &__content {
-    display: flex;
-    align-items: center;
-    height: 85%;
-
-    &__imageBox {
-      &__image {
-        max-width: 100%;
-        max-height: 100%;
-      }
-    }
-  }
-
-  &__footer {
-    height: 15%;
-    display: inline-block;
-    vertical-align: bottom;
-
-    &__title {
-      text-align: center;
-    }
-
-    &__description {
-      text-align: center;
-    }
-
-    &__category {
-      display: flex;
-      flex-direction: row;
-      justify-content: center;
-    }
-  }
-}
-
-.postButton {
-  width: 100%;
-  height: 100%;
-  border: none;
-  background-color: transparent;
-  position: relative;
-  &:focus {
-    outline: none;
-  }
-}
