+++ src/Pages/UserPage/index.scss
@@ -3,8 +3,6 @@
 */
 
 .userPage {
-  margin-top: 60px;
-
   &__contents {
     display: flex;
     flex-direction: column;
