+++ src/index.scss
@@ -5,7 +5,6 @@
 @import 'Components/Common/NavBar';
 @import 'Components/Form/TextField';
 @import 'Components/Post/Post';
-@import 'Components/Post/testPost';
 @import 'Components/Post/Comment';
 @import 'Components/Thumbnail';
 @import 'Components/SignUp/SignUpForm';
